Chore, Move, Docs:폴더구조 및 이름 변경, 리드미 변경, 탬플릿변경
