+++ .github/issue_template.md
@@ -1,3 +1,6 @@
 ## 목적
 
 ## 세부 내용
+
+## TODO-List
+- [ ]
