+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import HomePage from '@/containers/home/homePage';
+import HomePage from '@/templates/home/homePage';
 
 export default function Home() {
   return (
