+++ src/styles/app/home.scss
@@ -1,4 +1,4 @@
-@import 'globals';
+@import '@/styles/globals';
 
 .home {
   display: flex;
