+++ src/styles/app/login.scss
@@ -1,4 +1,4 @@
-@import 'globals';
+@import '@/styles/globals';
 
 .sub_header {
   font-size: 1.5rem;
