+++ src/styles/components/btn.scss
@@ -1,4 +1,4 @@
-@import 'globals';
+@import '@/styles/globals';
 
 button {
   background-color: $main-color;
